Add files via upload
This jupyter notebook contains the ability to select a book from the library and generate key characters and locations within the novel. Additionally, after selecting a chapter and filling in the narrator name The program will generate two types of questions from that chapter. It will additionally show the sentence the question so the user can under what the question is based off of and whether or not they want to use the question in their learning materials.